Move Collect proc-macro impl into a separate module
Also, un-hide the Collect macro reexport, so that the derive
documentation is visible directly from gc-arena's docs
