+++ src/gc-arena/src/collect.rs
@@ -12,10 +12,10 @@ use crate::context::CollectionContext;
 ///      [`Gc::write_barrier`](crate::Gc::write_barrier) during the same arena mutation.
 ///
 /// It is, however, possible to implement this trait safely by procedurally deriving it (see
-/// [`gc_arena_derive::Collect`]), which requires that every field in the structure also implement
-/// `Collect`, and implements a safe, empty version of `Drop`. Standard internally mutable types (like
-/// `std`'s `Cell` and `RefCell`) do not implement `Collect` in such a way that it is possible to store
-/// `Gc` pointers inside them, so the write barrier requirement cannot be broken when procedurally
+/// [`#[derive(Collect)]`](macro@crate::Collect)), which requires that every field in the structure also
+/// implement `Collect`, and implements a safe, empty version of `Drop`. Standard internally mutable
+/// types (like `std`'s `Cell` and `RefCell`) do not implement `Collect` in such a way that it is possible
+/// to store `Gc` pointers inside them, so the write barrier requirement cannot be broken when procedurally
 /// deriving `Collect`. A safe way of providing internal mutability in this case is to use the GC-aware
 /// cell types provided in the [`cell`](crate::cell) module, which allow internal mutability while
 /// ensuring that the write barrier is always executed.
